[aws] Reduce s3_handler memory usage for processing gzipped logs.
When reading `.gz` logs, to send back to datadog, multiple in-memory
copies of the decompressed data are being made.  This is ok in most cases, but when
the compressed files are already large (eg. 100MB+), the lambda function
tends to bomb out with Out of Memory exceptions.

Previously, the `s3_handler` would:

1. Read the entire contents of the gzipped file into memory from s3 (not too much choice here...),
2. Open a `gzip.GzipFile` iterator to walk each decompressed line of the file,
3. Immediately join all the lines ("all lines" list, unneeded copy 1) into one big `data` string (unneeded copy 2),
4. For the case of just "Send lines to Datadog", the `data` string is again split back in to a list of lines (unneeded copy 3).

This change tries to eliminate the copies created in steps 3 & 4.

In our local test case, memory usage against a 100MB gzip logfile dropped
from approx 3GB (and crashing...) down to 300MB (and completing...).

For the other cases catered for in this handler (`is_cloudtrail` and has a `DD_MULTILINE_LOG_REGEX_PATTERN`),
the behaviour remains the same as before.

For the `is_cloudtrail` case, it expects `data` to be a big blob of json to be parsed.
Whereas the multi-line regex check expects many lines in `data`...  May need to
rethink the approach for those if they prove problematic as well.