[aws] Convert DatadogBatcher.batch into a generator.
Similar to the previous few changes, the batch function was consuming
its given generator entirely before returning the full result list.

This change converts the routine into a generator, so it only yields a
batch at a time, instead of all the batches of all the logs.  This
prevents another potentially large list being created.