[aws] Removed early return from filter_logs.
Previously converted filter_logs into a generator function, to avoid
creating extra copies of the `logs` list.

However, since support for python2.x is expected, `return <value>` is illegal
inside a generator function.

I believe this check was to short-circuit the for loop, in
`filter_logs`, and skip checking for `EXCLUDE_AT_MATCH` and
`INCLUDE_AT_MATCH` for each log.  But I don't see the checking of these
variables against `None` to be costly.  The regex searches if they're
not `None` is the expensive part. So removing the short-circuit before
the for loop, in my mind, is ok.

Another option could be to replace it with a for-yield loop as well, but
I felt that might just be confusing.