[PROD] Update RDS func to deal with empty name tag
Currently filesystem metrics -> aws.rds.filesystem.* metrics are being tagged with a empty 'name:' tag. 
This happens because we are trying to parse the name key's value from the "fileSys" section of the log event, but it has an empty values as seen here: https://a.cl.ly/v1um7p1z which results in an empty tag like this: https://a.cl.ly/WnuEgNNn
I think my change will not add the `name` tag if it has an empty value.