Clarify Code Signing Documentation
The previous wording implied that associating a Code Signing Configuration with the Lambda function after deploying it would verify the integrity of the Forwarder. Because code signing checks only occur at deploy time, this is not the case. The only way to verify the integrity of the Forwarder is to use the manual installation method and associate a Code Signing Configuration with the Lambda function before uploading the ZIP file.